Added get movie list by genre API implementation in MainActivity.kt page.

+++ data/src/main/java/com/damai/data/responses/MovieListResponse.kt
@@ -31,7 +31,7 @@ class MovieItemResponse {
     var overview: String? = null
 
     @SerializedName("popularity")
-    var popularity: Int? = null
+    var popularity: Double? = null
 
     @SerializedName("poster_path")
     var posterPath: String? = null
